Proxy.Commands

Commands

attackMove :: UnitId -> Location -> Command

These are all the standard commands that you use

attackMove unit place This would make the unit unit move to place attacking enemy units enroute

rightClickAt :: UnitId -> Location -> Command

rightClickAt unit place will perform the action associated to right clicking at place on the map with unit unit.

patrol :: UnitId -> Location -> Command

patrol unit place will command unit to begin patrolling from its current location to that of place.

setRallyPosition :: UnitId -> Location -> Command

setRallyPosition unit place will set the rally point of unit to place.

land :: UnitId -> Location -> Command

land unit place will cause unit to land at place

unloadAllAt :: UnitId -> Location -> Command

unloadAllAt unit place will fully unload unit at place

attackUnit :: UnitId -> UnitId -> Command

First arg will attack second | ALL OF THE FOLLLOWING UNIT UNIT ARGUMENTS NEED TO BE CHECKED FOR ORDER

rightClickOn :: UnitId -> UnitId -> Command

Simulate a right click for arg1 on arg2

follow :: UnitId -> UnitId -> Command

arg1 follow arg2

setRallyUnit :: UnitId -> UnitId -> Command

I am not sure

repair :: UnitId -> UnitId -> Command

arg1 repair arg2

load :: UnitId -> UnitId -> Command

load arg2 into arg1

unload :: UnitId -> UnitId -> Command

unload arg2 from arg1

useTechOn :: UnitId -> TechType -> UnitId -> Command

arg1 use techType arg2 on unit arg3

train :: UnitId -> UnitType -> Command

arg1 train an instance of UnitType arg2

buildAddon :: UnitId -> UnitType -> Command

arg1 build addon structure arg2

morph :: UnitId -> UnitType -> Command

arg1 change into UnitType arg2

build :: UnitId -> Location -> UnitType -> Command

Use arg1 to build at arg2 unit arg3

useTechAt :: UnitId -> TechType -> Location -> Command

Use arg1 to perform tech arg2 at arg3

research :: UnitId -> TechType -> Command

Arg1 research tech arg2

useTech :: UnitId -> TechType -> Command

Arg1 use tech arg2

upgrade :: UnitId -> UpgradeType -> Command

Arg1 research upgrade arg2

stop :: UnitId -> Command

Arg1 stop your current action

holdPosition :: UnitId -> Command

Arg1 hold position. Different to stop, unit will not perform any action.

burrow :: UnitId -> Command

Arg1 burrow

unburrow :: UnitId -> Command

Arg1 unburrow

seige :: UnitId -> Command

Arg1 enter seige mode

unseige :: UnitId -> Command

Arg1 exit seige mode

cloak :: UnitId -> Command

Arg1 cloak

decloak :: UnitId -> Command

Arg1 decloak

lift :: UnitId -> Command

I am not sure

unloadAll :: UnitId -> Command

Arg1 unload all units being carried

cancelConstruction :: UnitId -> Command

Arg1 Cancel construction. I assume that this destroys the building and returns its resources

haltConstruction :: UnitId -> Command

Arg1 pause constructing the building

cancelMorph :: UnitId -> Command

Arg1 stop the morphing process

cancelTrain :: UnitId -> Command

Arg1 stop training the unit you are currently training

cancelAddon :: UnitId -> Command

Arg1 cancel building the addon you are building

cancelResearch :: UnitId -> Command

Arg1 cancel the research you are currently performing

cancelUpgrade :: UnitId -> Command

Arg1 cancel the upgrade you are currently researching

cancelTrainOn :: UnitId -> Slot -> Command

I am not sure

setGameSpeed :: Int -> Command

Set the game speed to be higher

Command preparation

Arguments are prepended to a string which contains 0's in the other arguement spaces delimited by ';'

s :: [Char] -> CommandArgs

Seperator for arguments

args1 :: Serialize a => a -> CommandArgs

Function to append a single argument

args2 :: (Serialize a, Serialize b) => a -> b -> CommandArgs

Function to append 2 argument

args3 :: (Serialize a, Serialize b, Serialize c) => a -> b -> c -> CommandArgs

Function to append 3 arguments

args4 :: (Serialize a, Serialize b, Serialize c, Serialize d) => a -> b -> c -> d -> CommandArgs

Function to specify all 4 arguments

Command Signatures

atLocation :: CommandType -> UnitId -> Location -> Command

atUnit :: CommandType -> UnitId -> UnitId -> Command

withTechTypeAtUnit :: CommandType -> UnitId -> TechType -> UnitId -> Command

atUnitType :: CommandType -> UnitId -> UnitType -> Command

atLocationWithUnitType :: CommandType -> UnitId -> Location -> UnitType -> Command

withTechTypeAtLocation :: CommandType -> UnitId -> TechType -> Location -> Command

atTechType :: CommandType -> UnitId -> TechType -> Command

atUpgradeType :: CommandType -> UnitId -> UpgradeType -> Command

atNothing :: CommandType -> UnitId -> Command

atSlot :: CommandType -> UnitId -> Slot -> Command

setValue :: CommandType -> Int -> Command