Proxy.Game

Documentation

type Id = Int

Identifier

type PlayerId = Id

Unique Player identifier

type UnitId = Id

Unique Unit identifier (not sure if this is unique per frame or per game)

type Name = String

Name

type Location = (Int, Int)

Location on map

type Width = Int

Width of object

type Height = Int

Height of object

type Slot = Int

Not sure

type Speed = Int

Speed of unit

type Buildable = Bool

Whether a unit is buildable

type Walkable = Bool

Not sure

data Tile

Contains information about a tile (consider changing to field name based)

Constructors

Tile Height Buildable Walkable

data Race

All the species in SC

Constructors

Zerg

Terran

Protoss

OtherRace

UnknownRace

SelectRace

RandomRace

NoneRace

data PlayerType

All player types

Constructors

NotUsed

Computer

Human

Rescuable

UnknownPlayerType0

ComputerSlot

OpenSlot

Neutral

ClosedSlot

UnknownPlayerType1

HumanDefeated

ComputerDefeated

data PlayerInfo

Information about a given player

Constructors

PlayerInfo

Fields

playerId :: PlayerId

playerRace :: Race

playerName :: Name

playerType :: PlayerType

data Player

DataTypes for each of the types of other players

Constructors

Me PlayerInfo

Ally PlayerInfo

Enemy PlayerInfo

data Map

Map information, name, dimensions and a list of areas,

Constructors

Map Name Width Height [[Tile]]

data Choke

Information about map regions which restrict movement, location and width

Constructors

Choke Location Width

type Minerals = Int

Number of minerals ?

type Gas = Int

Quanitity of Gas ?

data Resources

Resources of a player ?

Constructors

Resources Minerals Gas

type Health = Int

Unit Health

type Shields = Int

Unit Shields

type Energy = Int

Unit Energy

type Time = Int

Time

type OrderTimer = Int

Time until order completed?

type AddonId = Id

Number of Addons?

type MineCount = Int

Don't know

type Used = Int

Used to track total number of used unit supply

type Total = Int

Used to track current total of unit supply

data Supply

Player unit supply, used and total

Constructors

Supply Used Total

data TechStatus

The current status of a specific tech upgrade

Constructors

TechNotResearched

ResearchingTech

TechResearched

type Level = Int

Probably for current level of tech upgrade

data UpgradeStatus

Probably for unit special upgrades current status

Constructors

UpgradeNotResearched

ResearchingUpgrade

UpgradeResearched Level