Proxy.Messages

Documentation

type Ack = [Player]

I am not sure, a list of players

type StartingLocations = [Location]

List of starting locations, unsure if used for possible starting locations or for starting locations of players this match

data MapData

Map data, consider rewriting for field names

Constructors

MapData Name Width Height [Tile]

type ChokeData = [Choke]

List of Chokepoints on the map

type BaseData = [Location]

List of Base locations

data TimeData

Time data. Unclear what is used for at this stage

Constructors

TimeData

Fields

buildTime :: Time

trainTime :: Time

researchTime :: Time

upgradeTime :: Time

orderTimer :: OrderTimer

data UnitData

Used to carry data about an individual unit

Constructors

UnitData

Fields

unitId :: UnitId

unitOwnerId :: PlayerId

unitType :: UnitType

unitLocation :: Location

unitHealth :: Health

unitShields :: Shields

unitEnergy :: Energy

unitTimeData :: TimeData

unitOrder :: Order

unitResources :: Int

unitAddonId :: AddonId

unitMineCount :: MineCount

data GameState

Used to contain game state, probably of an individual player

Constructors

GameState

Fields

gameResources :: Resources

gameSupply :: Supply

gameTechStatus :: [(TechType, TechStatus)]

gameUpgradeStatus :: [(UpgradeType, UpgradeStatus)]

gameUnits :: [UnitData]

data GameInfo

Information about the current game

Constructors

GameInfo

Fields

gamePlayers :: [Player]

gameStartingLocations :: [Location]

gameMap :: MapData

gameTerrain :: Maybe (ChokeData, BaseData)

data Options

List of selected options about the current game

Constructors

Options [Bool]

type CommandArgs = String

String of arguments for a command

data Command

Data structure for a command

Constructors

Command CommandType CommandArgs

data Commands

A list of commands

Constructors

Commands [Command]

class Serialize a where

Methods

serializes :: a -> ShowS

serialize :: a -> String