Proxy.Settings

Description

some AI setup options including server port and history length

Documentation

historyLength :: Int

Number of previous game states to be passed to the next gameState

port :: PortID

Port to listen on for BWAPI connection

allowUserControl :: Bool

Allow the user to interact with the game alongside the AI

allowCompleteInformation :: Bool

Allow the AI access to all game data

printCommandsToConsole :: Bool

Print commands to console running the proxy

performTerrainAnalysis :: Bool

Perform terrain analysis at start