Proxy.Unit

Description

This module provides a number of functions to find details about a given unit type

Documentation

supplyRequired :: Num a => UnitType -> a

This returns the number of supplies required to field the unit

canAttack :: UnitType -> Bool

Returns if the unit is capable of attack

maxShields :: Num a => UnitType -> a

Returns the maximum shields of the unit

isBuilding :: UnitType -> Bool

Returns true if a unit is a building. (Some might be wrong)

isSpellCaster :: UnitType -> Bool

Returns if the unit type can cast special abilities

isFlyer :: UnitType -> Bool

Also wrong , should return if a unit is flying (Unclear if for terrain navigation or just not touching the floor)

buildTime :: Num a => UnitType -> a

Time (in seconds) required to build a unit

maxEnergy :: Num a => UnitType -> a

The maximum energy of a unit

isWorker :: UnitType -> Bool

Returns whether a unit can gather and build

airRange :: Num a => UnitType -> a

Attacking range in (Tiles or pixels not clear likely to be pixels based on sizes) of the unit against aerial foes

airDamage :: Num a => UnitType -> a

I think this should be the number of units of damage to air units an attack should do. This is problematic because of damage types http://starcraft.wikia.com/wiki/Damage_types consider

groundMinRange :: Num a => UnitType -> a

Minimum attacking distance (inpixels) of the unit to other ground based foes

groundMaxRange :: Num a => UnitType -> a

Maximum attacking distance of the unit to a ground based foe

tileHeight :: Num a => UnitType -> a

Height of unit in tiles

sightRange :: Num a => UnitType -> a

Sight range of unit in tiles

canMove :: UnitType -> Bool

Returns whether the unit can move

whatBuilds :: UnitType -> UnitType

Returns which unit builds the arguement unit

mineralsCost :: Num a => UnitType -> a

Returns the mineral cost to produce the unit

gasCost :: Num a => UnitType -> a

Returns the gas cost to produce the unit

maxHitPoints :: Num a => UnitType -> a

returns the number of units of damage a unit can recieve before being removed from the game

tileWidth :: Num a => UnitType -> a

returns the number of tiles occupied by the unit

groundDamage :: Num a => UnitType -> a

Returns the number of units of damage done by an attack from the unit

race :: UnitType -> Race

Returns the species of the given unit

supplyProvided :: Num a => UnitType -> a

returns the number of supplies returned by the unit for the player