Proxy.Upgrade

Description

This module contains the functions to associate a given upgrade with its cost and research location

Documentation

whatResearches :: UpgradeType -> UnitType

This function returns the building which allows the arguement upgrade.

The following functions are all used as part of the formula for calculating a given upgrade cost with the formula A in{Gas, Minerals} cost = ABase + Level*AFactor

gasBase :: Num a => UpgradeType -> a

This returns the number of gas units required for the first level of the upgrade ^ currently I think it is wrong.

mineralsFactor :: Num a => UpgradeType -> a

Returns the number of minerals the next level of the upgrade requires

gasFactor :: Num a => UpgradeType -> a

This returns the cost of the next level of an upgrade

mineralsBase :: Num a => UpgradeType -> a

Returns the cost of the first level of the given upgrade

repeats :: Num a => UpgradeType -> a

This function determines the number of levels this upgrade has